+++ emblem/src/main/scala/emblem/emblematic/traversors/sync/EmblematicToJsonTranslator.scala
@@ -18,38 +18,47 @@ import org.json4s.JsonAST.JObject
 import org.json4s.JsonAST.JString
 import org.json4s.JsonAST.JValue
 
-/** translates emblematic types into json4s AST */
+/** translates emblematic types into json4s AST.
+ * 
+ * non top-level emblems with a single property will be inlined in the JSON.
+ */
 class EmblematicToJsonTranslator extends Traversor {
 
-  type TraverseInput[A] = A
+  /** translates an emblematic type into json4s AST */
+  def translate[A : TypeKey](input: A): JValue = traverse[A](WrappedInput(input, true))
+
+  case class WrappedInput[A](value: A, isTopLevel: Boolean)
+  type TraverseInput[A] = WrappedInput[A]
   type TraverseResult[A] = JValue
 
-  override protected def traverseBoolean(input: Boolean): JValue = JBool(input)
+  override protected def traverseBoolean(input: WrappedInput[Boolean]): JValue = JBool(input.value)
 
-  override protected def traverseChar(input: Char): JValue = JString(input.toString)
+  override protected def traverseChar(input: WrappedInput[Char]): JValue = JString(input.value.toString)
 
-  override protected def traverseDateTime(input: DateTime): JValue = JString(dateTimeFormatter.print(input))
+  override protected def traverseDateTime(input: WrappedInput[DateTime]): JValue =
+    JString(dateTimeFormatter.print(input.value))
 
-  override protected def traverseDouble(input: Double): JValue = JDouble(input)
+  override protected def traverseDouble(input: WrappedInput[Double]): JValue = JDouble(input.value)
 
-  override protected def traverseFloat(input: Float): JValue = JDouble(input.toDouble)
+  override protected def traverseFloat(input: WrappedInput[Float]): JValue = JDouble(input.value.toDouble)
 
-  override protected def traverseInt(input: Int): JValue = JInt(input)
+  override protected def traverseInt(input: WrappedInput[Int]): JValue = JInt(input.value)
 
-  override protected def traverseLong(input: Long): JValue = JLong(input)
+  override protected def traverseLong(input: WrappedInput[Long]): JValue = JLong(input.value)
 
-  override protected def traverseString(input: String): JValue = JString(input)
+  override protected def traverseString(input: WrappedInput[String]): JValue = JString(input.value)
 
-  override protected def constituentTypeKey[A : TypeKey](union: Union[A], input: A): TypeKey[_ <: A] =
-    union.typeKeyForInstance(input).get
+  override protected def constituentTypeKey[A : TypeKey](union: Union[A], input: WrappedInput[A])
+  : TypeKey[_ <: A] =
+    union.typeKeyForInstance(input.value).get
 
-  override protected def stageUnion[A : TypeKey, B <: A : TypeKey](union: Union[A], input: A)
-  : Iterable[B] =
-    Seq(input.asInstanceOf[B])
+  override protected def stageUnion[A : TypeKey, B <: A : TypeKey](union: Union[A], input: WrappedInput[A])
+  : Iterable[WrappedInput[B]] =
+    Seq(WrappedInput(input.value.asInstanceOf[B], input.isTopLevel))
 
   override protected def unstageUnion[A : TypeKey, B <: A : TypeKey](
     union: Union[A],
-    input: A,
+    input: WrappedInput[A],
     result: Iterable[JValue])
   : JValue = {
     val fields = result.head.asInstanceOf[JObject].obj
@@ -58,25 +67,30 @@ class EmblematicToJsonTranslator extends Traversor {
 
   override protected def stageEmblemProps[A : TypeKey](
     emblem: Emblem[A],
-    input: A)
+    input: WrappedInput[A])
   : Iterable[PropInput[A, _]] = {
-    def propInput[B](prop: EmblemProp[A, B]) = prop -> prop.get(input)
+    def propInput[B](prop: EmblemProp[A, B]) = prop -> WrappedInput(prop.get(input.value), false)
     emblem.props.map(propInput(_))
   }
 
   override protected def unstageEmblemProps[A : TypeKey](
     emblem: Emblem[A],
+    input: WrappedInput[A],
     result: Iterable[PropResult[A, _]])
   : JValue = {
-    val jFields = result.toList.map { case (prop, result) => prop.name -> result }
-    JObject(jFields)
+    if (emblem.props.size == 1 && !input.isTopLevel) {
+      result.head._2
+    } else {
+      val jFields = result.toList.map { case (prop, result) => prop.name -> result }
+      JObject(jFields)
+    }
   }
 
   override protected def stageExtractor[Domain : TypeKey, Range : TypeKey](
     extractor: Extractor[Domain, Range],
     input: TraverseInput[Domain])
   : TraverseInput[Range] =
-    extractor.apply(input)
+    WrappedInput(extractor.apply(input.value), input.isTopLevel)
 
   override protected def unstageExtractor[Domain : TypeKey, Range : TypeKey](
     extractor: Extractor[Domain, Range],
@@ -87,29 +101,31 @@ class EmblematicToJsonTranslator extends Traversor {
   override protected def stageOptionValue[A : TypeKey](
     input: TraverseInput[Option[A]])
   : Iterable[TraverseInput[A]] =
-    input.toIterable
+    input.value.toIterable.map(WrappedInput(_, input.isTopLevel))
 
   override protected def unstageOptionValue[A : TypeKey](
-    input: Option[A],
+    input: WrappedInput[Option[A]],
     result: Iterable[JValue])
   : JValue =
     result.headOption.getOrElse(JNothing)
 
-  override protected def stageSetElements[A : TypeKey](input: Set[A]): Iterable[A] = input
+  override protected def stageSetElements[A : TypeKey](input: WrappedInput[Set[A]]): Iterable[WrappedInput[A]] =
+    input.value.map(WrappedInput(_, false))
 
   override protected def unstageSetElements[A : TypeKey](
-    input: TraverseInput[Set[A]],
+    input: WrappedInput[Set[A]],
     result: Iterable[JValue])
   : JValue =
     JArray(result.toList)
 
-  override protected def stageListElements[A : TypeKey](input: List[A]): Iterable[A] = input
+  override protected def stageListElements[A : TypeKey](input: WrappedInput[List[A]])
+  : Iterable[WrappedInput[A]] =
+    input.value.map(WrappedInput(_, false))
 
   override protected def unstageListElements[A : TypeKey](
-    input: List[A],
+    input: WrappedInput[List[A]],
     result: Iterable[JValue])
   : JValue =
     JArray(result.toList)
 
 }
-