+++ emblem/src/main/scala/emblem/emblematic/traversors/sync/Transformer.scala
@@ -124,6 +124,7 @@ trait Transformer {
 
     override protected def unstageEmblemProps[A : TypeKey](
       emblem: Emblem[A],
+      input: A,
       result: Iterable[PropResult[A, _]])
     : A = {
       val builder = emblem.builder()