+++ emblem/src/test/scala/emblem/emblematic/traversors/sync/EmblematicToJsonTranslatorSpec.scala
@@ -21,22 +21,22 @@ class EmblematicToJsonTranslatorSpec extends FlatSpec with GivenWhenThen with Ma
   behavior of "EmblematicToJsonTranslator.generate[A] for basic types"
 
   it should "produce the appropriate json4s values" in {
-    translator.traverse(true) should equal (JBool(true))
-    translator.traverse(false) should equal (JBool(false))
-    translator.traverse('q') should equal (JString("q"))
+    translator.translate(true) should equal (JBool(true))
+    translator.translate(false) should equal (JBool(false))
+    translator.translate('q') should equal (JString("q"))
     val dt = DateTime.now
-    translator.traverse(dt) should equal (JString(dateTimeFormatter.print(dt)))
-    translator.traverse(0.7d) should equal (JDouble(0.7d))
-    translator.traverse(0.7f) should equal (JDouble(0.7f))
-    translator.traverse(9) should equal (JInt(9))
-    translator.traverse(9L) should equal (JLong(9L))
-    translator.traverse("string") should equal (JString("string"))
+    translator.translate(dt) should equal (JString(dateTimeFormatter.print(dt)))
+    translator.translate(0.7d) should equal (JDouble(0.7d))
+    translator.translate(0.7f) should equal (JDouble(0.7f))
+    translator.translate(9) should equal (JInt(9))
+    translator.translate(9L) should equal (JLong(9L))
+    translator.translate("string") should equal (JString("string"))
   }
 
   behavior of "EmblematicToJsonTranslator.generate[Point]"
 
   it should "produce the appropriate json4s values" in {
-    translator.traverse(geometry.Point(0.4d, -0.3d)) should equal {
+    translator.translate(geometry.Point(0.4d, -0.3d)) should equal {
       JObject(List("x" -> JDouble(0.4d), "y" -> JDouble(-0.3d)))
     }
   }
@@ -44,7 +44,7 @@ class EmblematicToJsonTranslatorSpec extends FlatSpec with GivenWhenThen with Ma
   behavior of "EmblematicToJsonTranslator.generate[Polygon]"
 
   it should "produce the appropriate json4s values" in {
-    { translator.traverse(geometry.Polygon(Set(
+    { translator.translate(geometry.Polygon(Set(
         geometry.Point(0.44d, -0.34d),
         geometry.Point(0.45d, -0.35d),
         geometry.Point(0.46d, -0.36d))))