+++ src/main/scala/longevity/persistence/cassandra/JsonToPersistentTranslator.scala
@@ -10,7 +10,6 @@ import longevity.subdomain.Assoc
 import longevity.subdomain.AssocAny
 import longevity.subdomain.persistent.Persistent
 import org.json4s.JsonAST.JString
-import org.json4s.JsonAST.JValue
 
 private[cassandra] class JsonToPersistentTranslator(
   override protected val emblematic: Emblematic)
@@ -19,12 +18,12 @@ extends JsonToEmblematicTranslator {
   override protected val customTraversors = CustomTraversorPool.empty + assocTraversor
 
   def assocTraversor = new CustomTraversor[AssocAny] {
-    def apply[B <: Assoc[_ <: Persistent] : TypeKey](input: JValue): B = {
+    def apply[B <: Assoc[_ <: Persistent] : TypeKey](input: WrappedInput): B = {
       def assocFromString(s: String) = {
         val associateeTypeKey = typeKey[B].typeArgs(0).asInstanceOf[TypeKey[_ <: Persistent]]
         CassandraId(UUID.fromString(s))
       }
-      input match {
+      input.value match {
         case JString(s) => assocFromString(s).asInstanceOf[B]
         case _ => throw new CouldNotTraverseException(typeKey[B])
       }