+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [0.9-SNAPSHOT]
 
+- 2016.06.21 - single-property embeddables are now inlined. this has
+  no affect on the user other than how the persistents are translated
+  into JSON/BSON.
 - 2016.06.20 - add parent types `Embeddable` and `EType` for `Entity`,
   `ValueObject`, `EntityType`, and `ValueType`.
 - 2016.06.10 - add noop query `Query.All()`